lots of package upgrades, improvements to qol for importing jsx import emotion css prop, now built in and import is unecessary. refactored map card to imagecard component that will be reused for any selectable card instance.
