+++ package.json
@@ -11,38 +11,38 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@emotion/react": "^11.11.1",
+    "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.0",
-    "@fontsource/roboto": "^5.0.8",
+    "@fontsource/roboto": "^5.0.12",
     "@hookform/resolvers": "^3.3.4",
-    "@mui/icons-material": "^5.14.18",
-    "@mui/material": "^5.14.18",
-    "@mui/x-data-grid": "^6.18.2",
-    "@reduxjs/toolkit": "^1.9.7",
-    "axios": "^1.5.0",
-    "eslint-plugin-react": "^7.33.2",
+    "@mui/icons-material": "^5.15.14",
+    "@mui/material": "^5.15.14",
+    "@mui/x-data-grid": "^6.19.6",
+    "@reduxjs/toolkit": "^2.2.1",
+    "axios": "^1.6.8",
+    "eslint-plugin-react": "^7.34.1",
     "localforage": "^1.10.0",
-    "match-sorter": "^6.3.1",
+    "match-sorter": "^6.3.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-error-boundary": "^4.0.12",
-    "react-hook-form": "^7.50.1",
-    "react-redux": "^8.1.3",
-    "react-router-dom": "^6.16.0",
+    "react-error-boundary": "^4.0.13",
+    "react-hook-form": "^7.51.1",
+    "react-redux": "^9.1.0",
+    "react-router-dom": "^6.22.3",
     "sort-by": "^1.2.0",
-    "yup": "^1.3.3"
+    "yup": "^1.4.0"
   },
   "devDependencies": {
-    "@types/node": "^20.10.0",
-    "@types/react": "^18.2.15",
-    "@types/react-dom": "^18.2.7",
-    "@typescript-eslint/eslint-plugin": "^6.0.0",
-    "@typescript-eslint/parser": "^6.0.0",
-    "@vitejs/plugin-react": "^4.0.3",
-    "eslint": "^8.45.0",
+    "@types/node": "^20.11.30",
+    "@types/react": "^18.2.67",
+    "@types/react-dom": "^18.2.22",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
+    "@typescript-eslint/parser": "^7.3.1",
+    "@vitejs/plugin-react": "^4.2.1",
+    "eslint": "^8.57.0",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.3",
-    "typescript": "^5.0.2",
-    "vite": "^4.4.5"
+    "eslint-plugin-react-refresh": "^0.4.6",
+    "typescript": "^5.4.2",
+    "vite": "^5.1.6"
   }
 }
