+++ src/features/AppModal/AppModal.tsx
@@ -1,4 +1,3 @@
-/** @jsxImportSource @emotion/react */
 import { FC } from "react";
 import Backdrop from "@mui/material/Backdrop";
 import Box from "@mui/material/Box";
