+++ src/features/ButtonAppBar/ButtonAppBar.tsx
@@ -1,4 +1,3 @@
-/** @jsxImportSource @emotion/react */
 import AppBar from "@mui/material/AppBar";
 import Box from "@mui/material/Box";
 import Toolbar from "@mui/material/Toolbar";
