+++ src/features/MultiStepForm/forms/AddGameForm/SelectMap/SelectMap.tsx
@@ -1,57 +1,15 @@
-/** @jsxImportSource @emotion/react */
 import { FC, useCallback, useMemo } from "react";
 import { useFormContext } from "react-hook-form";
-import { Card, CardActionArea, CardMedia, CardContent, Typography, Grid, CircularProgress, useTheme } from "@mui/material";
+import { Typography, Grid, CircularProgress } from "@mui/material";
 import { useGetMapsQuery } from "../../../../../services/mapApi";
 import FilterButtons, { FilterOption } from "../../../../FilterButtons/FilterButtons";
 import { useAppSelector } from "../../../../../app/hooks";
 import { selectFilter } from "../../../../FilterButtons/filterButtonsSlice";
 import GameType, { GameTypes } from "../../../../../types/GameTypes.type";
-import OverwatchMap from "../../../../../types/OverwatchMap.type";
+import ImageCard from "../../../../../components/ImageCard/ImageCard";
 import snakeCase from "../../../../../utils/snakeCase";
-import useSelectMapStyles from "./useSelectMapStyles";
-
-type MapCardProps = {
-  map: OverwatchMap;
-  isSelected: boolean;
-  onCardClick: (id: number) => void;
-};
 
 type SelectMapProps = Record<string, unknown>;
-
-// refactor this into a re-usable card.
-const MapCard: FC<MapCardProps> = ({ map: { id, name }, isSelected, onCardClick }) => {
-  const theme = useTheme();
-  const classes = useSelectMapStyles();
-  const { constants: { mediaCardHeight } } = theme;
-
-  return (
-    <Card
-      onClick={() => onCardClick(id)} variant={"outlined"}
-      css={isSelected ? classes.cardIsSelected : undefined}
-    >
-      <CardActionArea>
-        <CardMedia
-          height={mediaCardHeight}
-          component={"img"}
-          alt={name}
-          image={`src/assets/maps/${snakeCase(name)}.webp`}
-        />
-        <CardContent>
-          <Typography
-            whiteSpace={"nowrap"}
-            overflow={"hidden"}
-            textOverflow={"ellipsis"}
-            gutterBottom
-            variant={"subtitle1"}
-          >
-            {name}
-          </Typography>
-        </CardContent>
-      </CardActionArea>
-    </Card>
-  );
-};
 
 const SelectMap: FC<SelectMapProps> = () => {
   const { setValue, watch, clearErrors } = useFormContext();
@@ -103,20 +61,25 @@ const SelectMap: FC<SelectMapProps> = () => {
         sx={{ maxHeight: 500, minHeight: 500, overflowY: "auto" }}
       >
         {!mapsLoading && selectedFilter &&
-          filteredMaps?.map((map) => (
-            <Grid
-              item
-              xs={12}
-              sm={6}
-              key={map.id}
-            >
-              <MapCard
-                map={map}
-                isSelected={selectedCard === map.id}
-                onCardClick={handleCardClick}
-              />
-            </Grid>
-          ))}
+          filteredMaps?.map((map) => {
+            const { name } = map;
+            return (
+              <Grid
+                item
+                xs={12}
+                sm={6}
+                key={map.id}
+              >
+                <ImageCard
+                  src={`src/assets/maps/${snakeCase(name)}.webp`}
+                  map={map}
+                  isSelected={selectedCard === map.id}
+                  onCardClick={handleCardClick}
+                />
+              </Grid>
+            );
+          }
+          )}
       </Grid>
     </Grid>
   );
