+++ src/layouts/Layout/Layout.tsx
@@ -1,4 +1,3 @@
-/** @jsxImportSource @emotion/react */
 import Box from "@mui/material/Box";
 import AppBar from "@mui/material/AppBar";
 import Toolbar from "@mui/material/Toolbar";
