+++ src/pages/LandingPage/LandingPage.page.tsx
@@ -1,4 +1,3 @@
-/** @jsxImportSource @emotion/react */
 import Button from "@mui/material/Button";
 import Paper from "@mui/material/Paper";
 import Box from "@mui/material/Box";
